test: the wrong-device path, and the broken message it was hiding
The audit moved the group-device check ahead of stage() so a wrong-device async
call could no longer poison a staging entry for the life of the group, and left
the case untested because it needs two GPUs. Both entry points are covered now,
plus the same shape working afterwards, which is the part the reordering is for.

The first run of it found the reason nothing had ever exercised that message:
c10::DeviceIndex is int8_t, so streaming it printed a CHARACTER rather than a
number. Device 0 is a NUL, which truncated the message at "input is on cuda:";
devices 1-3 emitted unprintable control bytes in place of the index. Both sites
cast to int64_t. It failed only on the rank whose stray device was 0, which is
why ws=3 passed and ws=4 did not.

The async destroyed-group path is covered too -- check_alive() has to run before
stage(), since destroy() returns early the second time and would never free an
entry created after it.

Two byte-capacity guards in prepare() stay untested, and validation.py's
docstring now says exactly why rather than leaving it to look like an oversight:
a smaller same-dtype buffer fails the shape check first, and a different-dtype
view of one needs storage torch will not hand out unless uneven splits make the
window strictly larger than this rank's output -- which succeeds on some ranks
and not others, and a rank-dependent `out` decides whether window() allocates,
so the worker would hang rather than report.

Verified on 4x B200 at ws=4 and ws=3, with the other six workers unchanged.

+++ src/bindings.cc
@@ -37,9 +37,11 @@ void require_cuda(const at::Tensor& input)
 // of the plan cache key.
 void require_group_device(const UlyssesGroup& group, const at::Tensor& input)
 {
+    // static_cast: c10::DeviceIndex is int8_t, so streaming it prints a CHARACTER. Device 0 is a
+    // NUL, which truncates the rest of the message.
     TORCH_CHECK(input.device().index() == group.device_index(),
                 "input is on cuda:",
-                input.device().index(),
+                static_cast<int64_t>(input.device().index()),
                 " but this UlyssesGroup was built for cuda:",
                 group.device_index(),
                 "; one group serves exactly one device");
@@ -92,7 +94,7 @@ Call prepare(const c10::intrusive_ptr<UlyssesGroup>&    group,
         TORCH_CHECK(call.output.is_cuda() && call.output.is_contiguous(), "out must be a contiguous CUDA tensor");
         TORCH_CHECK(call.output.device().index() == group->device_index(),
                     "out is on cuda:",
-                    call.output.device().index(),
+                    static_cast<int64_t>(call.output.device().index()),
                     " but this UlyssesGroup was built for cuda:",
                     group->device_index());
         TORCH_CHECK(call.output.scalar_type() == call.x.scalar_type(),
